requirements.txt
PyYAML
requests
tqdm
python-dotenv
gitpython
pandas
PyGithub
